refactor(src): define EmailLocale in src/types/emails, drop imports from emails/
Per PR review feedback — `src/` should not import from `emails/` for
primitive types. The directional rule is `emails/` may depend on
`src/`, never the reverse.

Add a new `sites/arolariu.ro/src/types/emails.ts` that mirrors only
the `EmailLocale` union and the `DEFAULT_EMAIL_LOCALE` constant.
The type is duplicated rather than re-exported from `emails/` so the
`src/` graph stays clean; a brief invariant comment in the file
explains that the union must stay in sync with the same-named type
in `emails/_lib/i18n.tsx` (a one-place rename would touch two files,
which is acceptable for this small surface).

Updated 3 imports:
- `src/lib/email/emailService.ts:13`
- `src/lib/actions/email/sendEmail.ts:18` (also updated the
  DEFAULT_LOCALE reference at L79 to DEFAULT_EMAIL_LOCALE)
- `src/app/domains/invoices/_dialogs/ShareInvoiceDialog.tsx:3`
