+++ sites/arolariu.ro/src/app/domains/invoices/_dialogs/ShareInvoiceDialog.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import type {EmailLocale} from "@/../emails/_lib/i18n";
+import type {EmailLocale} from "@/types/emails";
 import {sendEmail} from "@/lib/actions/email";
 import patchInvoice from "@/lib/actions/invoices/patchInvoice";
 import {LAST_GUID} from "@/lib/utils.generic";
