+++ sites/arolariu.ro/src/lib/actions/email/sendEmail.ts
@@ -15,7 +15,7 @@
 import {auth} from "@clerk/nextjs/server";
 import type {ReactElement} from "react";
 
-import {DEFAULT_LOCALE, type EmailLocale} from "@/../emails/_lib/i18n";
+import {DEFAULT_EMAIL_LOCALE, type EmailLocale} from "@/types/emails";
 import {emailTemplates, type EmailTemplateKey, type EmailTemplatePropsMap} from "@/../emails/_registry";
 
 // `emailService` is imported lazily inside the action body — NOT at the top
@@ -76,7 +76,7 @@ export async function sendEmail<K extends EmailTemplateKey>(input: SendEmailInpu
     return {success: false, error: `Unknown template: ${String(input.templateKey)}`};
   }
 
-  const locale: EmailLocale = input.props.locale ?? DEFAULT_LOCALE;
+  const locale: EmailLocale = input.props.locale ?? DEFAULT_EMAIL_LOCALE;
 
   // Narrow `variantProps` once — the discriminated registry union has
   // `variantProps` only on variant entries, but reading it through a
